+++ R/ParamDbl.R
@@ -1,6 +1,6 @@
 #' @rdname Domain
 #' @export
-p_dbl = function(lower = -Inf, upper = Inf, special_vals = list(), default = NO_DEF, tags = character(), tolerance = sqrt(.Machine$double.eps), depends = NULL, trafo = NULL, logscale = FALSE, init) {
+p_dbl = function(lower = -Inf, upper = Inf, special_vals = list(), default = NO_DEF, tags = character(), tolerance = sqrt(.Machine$double.eps), depends = NULL, trafo = NULL, logscale = FALSE, init, aggr = NULL) {
   assert_number(tolerance, lower = 0)
   assert_number(lower)
   assert_number(upper)
@@ -18,7 +18,7 @@ p_dbl = function(lower = -Inf, upper = Inf, special_vals = list(), default = NO_
   }
 
   Domain(cls = "ParamDbl", grouping = "ParamDbl", lower = real_lower, upper = real_upper, special_vals = special_vals, default = default, tags = tags, tolerance = tolerance, trafo = trafo, storage_type = "numeric",
-    depends_expr = substitute(depends), init = init, cargo = if (logscale) "logscale")
+    depends_expr = substitute(depends), init = init, cargo = if (logscale) "logscale", aggr = aggr)
 }
 
 #' @export