+++ R/ParamUty.R
@@ -1,7 +1,7 @@
 
 #' @rdname Domain
 #' @export
-p_uty = function(custom_check = NULL, special_vals = list(), default = NO_DEF, tags = character(), depends = NULL, trafo = NULL, repr = substitute(default), init) {
+p_uty = function(custom_check = NULL, special_vals = list(), default = NO_DEF, tags = character(), depends = NULL, trafo = NULL, repr = substitute(default), init, aggr = NULL) {
   assert_function(custom_check, null.ok = TRUE)
   if (!is.null(custom_check)) {
     custom_check_result = custom_check(1)
@@ -12,7 +12,7 @@ p_uty = function(custom_check = NULL, special_vals = list(), default = NO_DEF, t
   } else {
     "NoDefault"
   }
-  Domain(cls = "ParamUty", grouping = "ParamUty", cargo = list(custom_check = custom_check, repr = repr), special_vals = special_vals, default = default, tags = tags, trafo = trafo, storage_type = "list", depends_expr = substitute(depends), init = init)
+  Domain(cls = "ParamUty", grouping = "ParamUty", cargo = list(custom_check = custom_check, repr = repr), special_vals = special_vals, default = default, tags = tags, trafo = trafo, storage_type = "list", depends_expr = substitute(depends), init = init, aggr = aggr)
 }
 
 #' @export